Part of DataLoader v2 hackathon, this PR implements a new DataPipe that can read Parquet files and return TorchArrow DataFrames for each row group within the Parquet files.

CI failures of TorchVision is unrelated.

This PR is dependent on a change from TorchArrow

Differential Revision: [D33803733](https://our.internmc.facebook.com/intern/diff/D33803733)

[ghstack-poisoned]

+++ torchdata/datapipes/iter/util/dataframemaker.py
@@ -90,4 +90,4 @@ class ParquetDFLoaderIterDataPipe(IterDataPipe):  # IterDataPipe[torcharrow.IDat
             for i in range(num_row_groups):
                 # TODO: More fine-grain control over the number of rows or row group per DataFrame
                 row_group = parquet_file.read_row_group(i, columns=self.columns, use_threads=self.use_threads)
-                yield torcharrow.from_arrow(row_group)
+                yield torcharrow.from_arrow(row_group, dtype=self.dtype)
